nvptx executable
