+++ paddle/fluid/compiler/piano/backends/llvm_ir/CMakeLists.txt
@@ -17,8 +17,11 @@ cc_test(nvptx_ir_emitter_test SRCS nvptx_ir_emitter_test.cc
         DEPS nvptx_ir_emitter note_ir note_proto)
 target_link_libraries(nvptx_ir_emitter_test ${LLVM_LIBS})
 
+cc_library(executable_pool SRCS executable_pool.cc DEPS dynload_cuda)
+cc_library(nvptx_executable SRCS nvptx_executable.cc DEPS dynload_cuda)
 cc_library(llvm_compiler SRCS llvm_compiler.cc DEPS nvptx_ir_emitter nvptx_primitive_ir_emitter note_ir llvm)
-cc_library(nvptx_compiler SRCS nvptx_compiler.cc DEPS llvm_compiler llvm dynload_cuda)
+cc_library(nvptx_compiler SRCS nvptx_compiler.cc DEPS llvm_compiler llvm)
 cc_test(nvptx_compiler_test SRCS nvptx_compiler_test.cc
-        DEPS nvptx_compiler note_ir note_builder)
+        DEPS nvptx_compiler note_ir note_builder
+        nvptx_executable executable_pool)
 target_link_libraries(nvptx_compiler_test ${LLVM_LIBS})
